requirements.txt
Flask==3.0.0
python-dotenv==1.1.0
Werkzeug==3.0.0
blinker==1.9.0
itsdangerous==2.1.2
